+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `node-auto-repair` feature to automaticaly terminate bad and unhealthy nodes.
+- Add `terminate-unhealthy-node` feature to automaticaly terminate bad and unhealthy nodes in a Cluster.
 
 ## [9.1.3] - 2020-10-21